requirements.txt
beaker-pyteal==0.4.1
mypy==0.910
py-algorand-sdk==1.20.2
pyteal==0.20.1
pytest
